requirements.txt
discord.py==1.7.3
discord-py-interactions==3.0.2
yt-dlp
